moved specs to test folder
