+++ zio-webapp-workshop/src/main/scala/webapp/workshop/10-graduation.scala
@@ -1,6 +1,6 @@
 package webapp.workshop
 
-object Graduation {
+object GraduationSection {
 
   /**
    * EXERCISE
